requirements.txt.txt
Flask
flask-cors
google-generativeai